requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.3.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.2.0
google-generativeai==0.8.4
pypdf==4.3.1
requests==2.32.0
aiofiles==24.1.0
python-multipart==0.0.7
